reformated the code. added react router. got getUserById in its own file
